+++ include/editormode.hpp
@@ -1,0 +1,335 @@
+#pragma once
+
+#include "gridmap.hpp"
+
+class UndoStack
+{
+public:
+    UndoStack(s32 stackLimit = -1)
+        : mStackLimit(stackLimit)
+    {
+
+    }
+
+    template<class T, class... Args>
+    void PushMerge(bool shouldMerge, Args&&... args)
+    {
+        // If the last item on the stack is of the same type we are about to create
+        if (shouldMerge && !mUndoStack.empty() && GenerateTypeId<T>() == mUndoStack.back()->Id() && mUndoStack.back()->CanMerge())
+        {
+            // Undo the last action
+            mUndoStack.back()->Undo();
+
+            // Update its internals with the new target data
+            static_cast<T*>(mUndoStack.back().get())->Merge(std::forward<Args>(args)...);
+
+            // Re-apply with new internals
+            mUndoStack.back()->Redo();
+        }
+        else
+        {
+            Push<T>(std::forward<Args>(args)...);
+        }
+    }
+
+    template<class T, class... Args>
+    void Push(Args&&... args)
+    {
+        // If the active index isn't the latest item then remove everything after it
+        if (mCommandIndex != Count())
+        {
+            mUndoStack.erase(mUndoStack.begin() + mCommandIndex, mUndoStack.end());
+        }
+
+        // Apply action of new command and add to stack
+        auto cmd = std::make_unique<T>(std::forward<Args>(args)...);
+        mUndoStack.emplace_back(std::move(cmd));
+        mUndoStack.back()->Redo();
+
+        // If we are over the stack limit remove the first item to stay within the limit
+        if (mStackLimit != -1 && static_cast<s32>(Count()) > mStackLimit)
+        {
+            mUndoStack.erase(mUndoStack.begin(), mUndoStack.begin() + 1);
+        }
+        else
+        {
+            mCommandIndex++;
+        }
+    }
+
+    void Undo();
+    void Redo();
+    void Clear();
+    u32 Count() const { return static_cast<u32>(mUndoStack.size()); }
+    void DebugRenderCommandList(GuiContext& gui) const;
+private:
+    std::vector<std::unique_ptr<class ICommand>> mUndoStack;
+    u32 mCommandIndex = 0;
+    s32 mStackLimit = -1;
+};
+
+
+u32 NextId();
+
+template<class T>
+u32 GenerateTypeId()
+{
+    static u32 id = NextId();
+    return id;
+}
+
+
+class ICommand
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(ICommand);
+    ICommand() = default;
+    virtual void Redo() = 0;
+    virtual void Undo() = 0;
+    virtual std::string Message() = 0;
+    virtual ~ICommand() = default;
+    virtual u32 Id() const = 0;
+    virtual bool CanMerge() const { return false; }
+};
+
+template<class T>
+class ICommandWithId : public ICommand
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(ICommandWithId);
+    ICommandWithId()
+    {
+        mId = GenerateTypeId<T>();
+    }
+    virtual u32 Id() const override { return mId; }
+private:
+    u32 mId;
+};
+
+class Selection
+{
+public:
+    std::set<s32> Clear(CollisionLines& items);
+    void Select(CollisionLines& items, s32 idx, bool select);
+    bool HasSelection() const { return mSelectedLines.empty() == false; }
+    const std::set<s32>& SelectedLines() const { return mSelectedLines; }
+private:
+    std::set<s32> mSelectedLines;
+};
+
+class CommandSelectOrDeselectLine : public ICommandWithId<CommandSelectOrDeselectLine>
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(CommandSelectOrDeselectLine);
+
+    CommandSelectOrDeselectLine(CollisionLines& lines, Selection& selection, s32 idx, bool select)
+        : mLines(lines), mSelection(selection), mIdx(idx), mSelect(select) { }
+
+    virtual void Redo() final
+    {
+        mSelection.Select(mLines, mIdx, mSelect);
+    }
+
+    virtual void Undo() final
+    {
+        mSelection.Select(mLines, mIdx, !mSelect);
+    }
+
+    virtual std::string Message() final
+    {
+        if (mSelect)
+        {
+            return "Select line " + std::to_string(mIdx);
+        }
+        else
+        {
+            return "De select line " + std::to_string(mIdx);
+        }
+    }
+private:
+    CollisionLines& mLines;
+    Selection& mSelection;
+    const s32 mIdx;
+    const bool mSelect;
+};
+
+class CommandClearSelection : public ICommandWithId<CommandClearSelection>
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(CommandClearSelection);
+
+    CommandClearSelection(CollisionLines& lines, Selection& selection)
+        : mLines(lines), mSelection(selection) { }
+
+    virtual void Redo() final
+    {
+        mOldSelection = mSelection.Clear(mLines);
+    }
+
+    virtual void Undo() final
+    {
+        for (s32 idx : mOldSelection)
+        {
+            mSelection.Select(mLines, idx, true);
+        }
+    }
+
+    virtual std::string Message() final
+    {
+        return "Clear selection";
+    }
+
+private:
+    CollisionLines& mLines;
+    Selection& mSelection;
+    std::set<s32> mOldSelection;
+};
+
+inline std::string FormatVec2(const glm::vec2& vec)
+{
+    std::ostringstream out;
+    out << "("
+        << static_cast<s32>(vec.x)
+        << ","
+        << static_cast<s32>(vec.y)
+        << ")";
+    return out.str();
+}
+
+class CommandMoveLinePoint : public ICommandWithId<CommandMoveLinePoint>
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(CommandMoveLinePoint);
+
+    CommandMoveLinePoint(CollisionLines& lines, Selection& selection, const glm::vec2& newPos, bool moveP1)
+        : mLines(lines), mSelection(selection), mNewPos(newPos), mApplyToP1(moveP1)
+    {
+        mOldPos = Point();
+    }
+
+    void Merge(CollisionLines& /*lines*/, Selection& /*selection*/, const glm::vec2& newPos, bool /*moveP1*/)
+    {
+        mNewPos = newPos;
+    }
+
+    virtual void Redo() final
+    {
+        Point() = mNewPos;
+    }
+
+    virtual void Undo() final
+    {
+        Point() = mOldPos;
+    }
+
+    virtual bool CanMerge() const final { return true; }
+
+    virtual std::string Message() final
+    {
+        return "Move line point " + std::to_string(mApplyToP1 ? 1 : 2) + " from " + FormatVec2(mOldPos) + " to " + FormatVec2(mNewPos);
+    }
+private:
+    glm::vec2& Point()
+    {
+        const auto& line = mLines[*mSelection.SelectedLines().begin()];
+        return mApplyToP1 ? line->mLine.mP1 : line->mLine.mP2;
+    }
+
+    CollisionLines& mLines;
+    Selection& mSelection;
+    glm::vec2 mOldPos;
+    glm::vec2 mNewPos;
+    bool mApplyToP1;
+};
+
+
+class MoveSelection : public ICommandWithId<MoveSelection>
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(MoveSelection);
+
+    MoveSelection(CollisionLines& lines, Selection& selection, const glm::vec2& delta)
+        : mLines(lines), mSelection(selection), mDelta(delta)
+    {
+
+    }
+
+    void Merge(CollisionLines& /*lines*/, Selection& /*selection*/, const glm::vec2& delta)
+    {
+        mDelta += delta;
+    }
+
+    virtual void Redo() final
+    {
+        for (s32 idx : mSelection.SelectedLines())
+        {
+            mLines[idx]->mLine.mP1 += mDelta;
+            mLines[idx]->mLine.mP2 += mDelta;
+        }
+    }
+
+    virtual void Undo() final
+    {
+        for (s32 idx : mSelection.SelectedLines())
+        {
+            mLines[idx]->mLine.mP1 -= mDelta;
+            mLines[idx]->mLine.mP2 -= mDelta;
+        }
+    }
+
+    virtual std::string Message() final
+    {
+        return "Move selection by " + FormatVec2(mDelta);
+    }
+
+    virtual bool CanMerge() const final
+    {
+        return true;
+    }
+private:
+    CollisionLines& mLines;
+    Selection& mSelection;
+    glm::vec2 mDelta;
+};
+
+class EditorMode
+{
+public:
+    NO_MOVE_OR_MOVE_ASSIGN(EditorMode);
+
+    EditorMode(GridMapState& mapState);
+   
+
+    /* TODO: Set correct cursors
+    enum class eMouseCursor
+    {
+    eArrow,
+    eOpenHand,
+    eClosedHand,
+    };
+    eMouseCursor mMouseCursor = eMouseCursor::eArrow;
+    */
+
+    void Update(const InputState& input, CoordinateSpace& coords);
+    void Render(Renderer& rend, GuiContext& gui) const;
+
+    f32 mEditorCamZoom = 1.0f;
+
+private:
+
+    Selection mSelection;
+    UndoStack mUndoStack;
+    bool mMergeCommand = false;
+    glm::vec2 mLastMousePos;
+
+    enum class eSelectionState
+    {
+        eNone,
+        eLineP1Selected,
+        eLineP2Selected,
+        eLineMiddleSelected,
+        eMoveSelected
+    };
+    eSelectionState mSelectionState = eSelectionState::eNone;
+    GridMapState& mMapState;
+};
